Add inplace upgrader controller to hcco
This PR distributes the inplace upgrade logic from the NodePool controller into individual controllers running for each control plane namespace.
This let us honour our intentional network design restrictions to access hosted clusters and scope the impact and resources for each controller to specific hosted clusters.
The implementation at the moment uses annotations on MachineSets to communicate upgrade progress between NodePool controller (NodePool <-> MachineSet) and inplace upgrader (MachineSet <-> Nodes) i.e NodePool <-> Nodes. This info is modeled in the nodePoolUpgradeAPI struct which in future might be exposed as a CRD if needed.
The controller is added to the hcco manager as it already solves the problem of watching and caching targeted resources in both management and hosted cluster.
