+++ control-plane-operator/hostedclusterconfigoperator/controllers/inplaceupgrader/inplaceupgrader_test.go
@@ -1,4 +1,4 @@
-package nodepool
+package inplaceupgrader
 
 import (
 	"context"
