+++ hypershift-operator/controllers/nodepool/manifests.go
@@ -25,10 +25,9 @@ func machineDeployment(nodePool *hyperv1.NodePool, clusterName string, controlPl
 }
 
 func machineSet(nodePool *hyperv1.NodePool, clusterName string, controlPlaneNamespace string) *capiv1.MachineSet {
-	resourcesName := generateName(clusterName, nodePool.Spec.ClusterName, nodePool.GetName())
 	return &capiv1.MachineSet{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      resourcesName,
+			Name:      nodePool.GetName(),
 			Namespace: controlPlaneNamespace,
 		},
 	}
@@ -40,32 +39,6 @@ func machineHealthCheck(nodePool *hyperv1.NodePool, controlPlaneNamespace string
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      nodePool.GetName(),
 			Namespace: controlPlaneNamespace,
-		},
-	}
-}
-
-func inPlaceUpgradePod(namespace, nodeName string) *corev1.Pod {
-	return &corev1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: namespace,
-			Name:      fmt.Sprintf("machine-config-daemon-%s", nodeName),
-		},
-	}
-}
-
-func inPlaceUpgradeNamespace(nodePool *hyperv1.NodePool) *corev1.Namespace {
-	return &corev1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: fmt.Sprintf("%s-upgrade", nodePool.GetName()),
-		},
-	}
-}
-
-func inPlaceUpgradeConfigMap(nodePool *hyperv1.NodePool, namespace string) *corev1.ConfigMap {
-	return &corev1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: namespace,
-			Name:      fmt.Sprintf("%s-upgrade", nodePool.GetName()),
 		},
 	}
 }
